# MAINT: Improve performance of isscalar