@superbobry 2019-07-15
Thanks for the feedback! 

I agree that using `ndim(...) == 0` (generally) makes more sense, but the current implementation of `ndim` is only fast for NumPy scalars and arrays. For any other object it will attempt to convert its argument to an array (incurring a copy) and then return `ndim` of the result. 

Here're the timings:

```
>>> o = object()
>>> %timeit np.ndim(42.0) == 0
1.08 µs ± 29.7 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
>>> %timeit np.ndim([]) == 0
1.51 µs ± 59.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
>>> %timeit np.ndim(o) == 0
2.52 µs ± 97.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
 ```

I think a reasonable expectation from `ndim` is that it does not allocate and works in constant time. Making it such while preserving the current semantics (works for ~anything) could be tricky.

As an aside, `isscalar` also does not handle `memoryview` correctly since `memoryview` is listed in `np.core.numerictypes.ScalarType` (!).

```
>>> np.isscalar(memoryview(np.arange(10)))                                                                                                                                                                                                                                            
True
```