@eric-wieser 2019-07-16
> As an aside, isscalar also does not handle memoryview correctly

"correctly" for `isscalar` means "behaves as it used to in very old versions of numpy". The non-broken semantics are those of `ndim`.

Adding a fast-path for `np.ndim` on builtin obviously-scalar types would be a pretty reasonable idea.