@mattip 2019-08-08
Closing, please reopen if I misunderstood. @superbobry would you like to tackle adding a fast-path for `np.ndim` on scalar types ?